feat(monitor): merge Bundesländer tab into Umfragen
Selecting a Land in the existing 'Grüne in den Ländern' grid now also
shows the latest Landtagswahl results and the state-level Meinungsbild
(replacing the national one), so the separate Bundesländer tab and its
Combobox selector are removed. The Combobox was unusable anyway because
max-w-sm resolves to a 12px spacing token in this app.

BundeslandView.tsx becomes LandDetails.tsx, keeping only the reusable
sections (LandtagsergebnisCard, MeinungsbildForState, GerdaAttribution).
PARTY_COLORS moves to partyColors.ts to avoid an import cycle.
